ui : Add Qrcode interface improvements
